requirements.txt
streamlit>=1.31.0
boto3>=1.37.0
python-decouple>=3.8
importlib-metadata<7.0,>=1.4.0
scikit-learn==1.5.2
nltk==3.9.1
uuid==1.30